refactor: 유효성 검사 위치 수정
Before : Controller에서 Validator클래스들을 이용하여 유효성 검사

After : 각 모델을 생성할 떄 유효성 검사를 실시하도록 위임해 객체의 일관성을 보장할 수 있게 함
