+++ README.md
@@ -10,7 +10,7 @@
     5. [우승자 출력](#5-사용자에게-우승자-출력하기)
 
 - [추가 사항](#추가-사항)
-    1. [예외 처리](#예외처리)
+    1. [예외 처리](#-예외처리)
     2. [리팩토링 계획](#리펙토링)
 
 # 📄기능 구현 목록
@@ -128,4 +128,4 @@
 - [x] 일급 컬렉션 사용을 위해 `Racing` 클래스 안의 `List<Car> cars` 필드 분리
 - [x] `OutputView`에 비즈니스 로직의 결과를 DTO를 통해 한번에 전달
     - 기존 로직은 `System.out.print`를 비즈니스 로직 중간중간에 호출하여 I/O비용이 큼
-- [ ] 유효성 검사 실시 시점에 대해 기준 마련
+- [x] 유효성 검사 실시 시점에 대해 기준 마련
